pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "oha"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "gradio (>=5.29.0,<6.0.0)",
    "langchain (>=0.3.24,<0.4.0)",
    "langchain-core (>=0.3.57,<0.4.0)",
    "langchain-ollama (>=0.3.2,<0.4.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
